requirements.txt
opencv-python
moviepy
openai
whisper
torch
python-dotenv
